[SYSTEMDS-3363] Netty Encoding Buffer Allocation
This commit replaces the ObjectEncoders for encoding federated
communication messages (i.e., FederatedRequest and FederatedResponse)
with a respective FederatedRequestEncoder and FederatedResponseEncoder.
These two new classes extent from ObjectEncoder and override the
allocateBuffer method in order to allocate the buffer with an initial
capacity corresponding to the FederatedRequest/FederatedResponse instead
of the default value of 256.
I've tried estimating the entire serialized size of the requests and
responses to allocate the exact size needed, but haven't found a good
solution for it because of the many cases to consider. Hence, the initial
capacity in this PR only results from taking into consideration a static
offset (FederatedRequest: 512, FederatedResponse: 312), the size of the
CacheBlock parameters, and the length of the lineage trace (only for
FederatedRequests).
In my experiments with a matrix of dimensions 600.000x200, it reduces
the encoding time from 50 seconds to 1.9 seconds